test(backend): satisfy editor request CI gates
